requirements-dev.txt
-r requirements.txt
pytest>=7.3
pytest-cov>=4.1
pytest-asyncio>=0.21
pytest-httpx>=0.22
mypy>=1.3
ruff>=0.1.0
pre-commit>=3.3
hatchling